Fix "The input line is too long." when building on VS2013
If you get 'The input line is too long.' and 'The syntax of the command is incorrect.' messages, then vcvars32.bat has been run too many times. Close the Command Window, open a new one and it will work.

Put a message immediately before, and a comment in the script.