Updates and bug fixes
Add support for building with additional editions of VS 2017 and VS 2019
Update the readme file
Deploy script now checks that output is present before attempting to deploy and issues a better message if it isn't.
By customer request, the project now uses the latest versions of the AWS SDK Core and GameLift assemblies. (I don't anticipate really keeping these up to date unless requested, so be sure to open an issue if you want me to do this again)
Fix a bug that I introduced that caused me to make invalid service role policies for getting builds from S3. Note that this means that if you have a Role called GameLiftS3Access in your account, this can be deleted along with the attached policy, completely safely.